Styling and OnBlur removal
Added CSS styling, removed the OnBlur error styling and messages. Created a Button primitive as well but needs approval.
