+++ dwellingly/src/scss/main.scss
@@ -16,7 +16,7 @@
 // @import '../node_modules/bulma/sass/layout/section.sass';
 
 //@import '../../node_modules/bulma/sass/elements/_all.sass';
-//@import '../../node_modules/bulma/sass/grid/_all.sass';
+@import '../../node_modules/bulma/sass/grid/_all.sass';
 
 @import './helpers.scss';
 @import './typography.scss';
